refactor(core): classes
Move classes in a dedicated directory + update jsdoc
